requirements.txt
celery==5.4.0
colorama==0.4.6
Flask==3.1.0
itsdangerous==2.2.0
pydantic==2.10.4
PyJWT==2.10.1
pymongo==4.10.1
python-dotenv==1.0.1
python_bcrypt==0.3.2
redis==5.2.1
Requests==2.32.3
